requirements.txt
Flask==2.1.2
flask-jwt-extended==4.4.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
flask-reuploaded==1.2.0
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.28.0
passlib==1.7.4
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.36
sqlalchemy-utils==0.38.2